# Vectorize more of asymptotic expansion