As discussed in #44 there is some room for improved polynomial evaluation using simd.

There is also a big difference in evaluation compared to what instruction set you have.

On an linux with an Intel CPU
```julia
julia> @benchmark Bessels.besseljy_large_argument(v, x) setup=(v = rand()+20.0; x=rand()+30.0)
BenchmarkTools.Trial: 10000 samples with 893 evaluations.
 Range (min … max):  126.795 ns … 161.896 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     128.867 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   128.566 ns ±   1.552 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▆█             ▅▃                                             
  ▃██▇▅▅▄▄▄▃▂▂▂▂▃▄███▆▄▃▅▅▃▅▆▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▁▂▂▂▂ ▃
  127 ns           Histogram: frequency by time          135 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels.besseljy_large_argument(v, x) setup=(v = rand()+20.0; x=rand()+50.0)
BenchmarkTools.Trial: 10000 samples with 969 evaluations.
 Range (min … max):  77.175 ns … 118.102 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     77.439 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   77.805 ns ±   2.531 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█▃     ▂▃▁                                                  ▂
  ███▅▅▅▅▁█████▇▆▅▆▇▆▇▇▇▆▆▅▅▆▅▅▅▄▄▅▄▅▃▁▁▃▃▄▃▃▄▅▅▄▃▄▁▁▄▁▁▁▄▃▁▁▃ █
  77.2 ns       Histogram: log(frequency) by time      87.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels.besseljy_large_argument(v, x) setup=(v = rand()+20.0; x=rand()+150.0)
BenchmarkTools.Trial: 10000 samples with 985 evaluations.
 Range (min … max):  54.440 ns … 84.694 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     54.930 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   55.131 ns ±  1.355 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▅▅█▄      ▁▂▂                                              ▁
  █████▄▁▃▃▄▅████▇▇▆▅▆▆▅▆▆▆▅▅▄▇▇▅▃▄▄▃▄▅▁▁▃▁▃▃▁▃▃▁▁▃▁▃▁▁▁▃▄▁▁▃ █
  54.4 ns      Histogram: log(frequency) by time      62.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

and on the M1 MacOS 
```julia
julia> @benchmark Bessels.besseljy_large_argument(v, x) setup=(v= rand() + 20.0; x = rand()+30.0)
BenchmarkTools.Trial: 10000 samples with 788 evaluations.
 Range (min … max):  159.846 ns … 236.463 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     160.162 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   160.683 ns ±   2.721 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▆█                                                          
  ▂▄██▇▃▂▂▂▂▂▂▂▂▂▂▂▁▁▂▁▂▂▂▂▁▂▂▁▂▁▂▂▂▂▃▃▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▂
  160 ns           Histogram: frequency by time          166 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels.besseljy_large_argument(v, x) setup=(v= rand() + 20.0; x = rand()+50.0)
BenchmarkTools.Trial: 10000 samples with 875 evaluations.
 Range (min … max):  134.238 ns … 196.666 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     134.666 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   135.196 ns ±   2.537 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂▇█▇▅    ▃          ▃▃▂                                       ▂
  █████▇▅▅██▇▅▆▆▆▇▄▅▅▇███▆▆▄▄▇▆▄▅▆▆▆▇▇▇▄▃▁▁▃▁▁▃▃▄▄▃▄▄▆▅▅▆▄▅▅▅▅▄ █
  134 ns        Histogram: log(frequency) by time        144 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels.besseljy_large_argument(v, x) setup=(v= rand() + 20.0; x = rand()+150.0)
BenchmarkTools.Trial: 10000 samples with 973 evaluations.
 Range (min … max):  73.826 ns … 135.064 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     74.084 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   74.773 ns ±   2.926 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▇        ▃▃                                                 ▁
  ███▃▂▇▆▅▄▄██▅▅▄▆▆▇▅▄▄▃▅▅▃▄▅▄▅▄▄▄▃▄▄▄▆▅▅▅▆▆▆▄▃▄▅▄▅▄▅▅▅▅▄▄▄▄▄▅ █
  73.8 ns       Histogram: log(frequency) by time      90.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

So we can see that there is quite the speed up by having the `<4 x>` instruction set indicating that we already were getting some vectorization on master.

This PR improves a little bit those performances.....

```julia
# master
julia> @benchmark Bessels._α_αp_asymptotic(v, x) setup=(v = rand()+20.0; x=rand()+150.0)
BenchmarkTools.Trial: 10000 samples with 995 evaluations.
 Range (min … max):  28.868 ns … 41.187 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     29.333 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   29.356 ns ±  0.364 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄             █▁                                        ▁   ▁
  █▅▄▃▁▃▃▄▃▃▁▁▃▁███▇▄▆▅▄▅▁▃▃▃▁▃▃▁▁▁▁▁▃▁▁▁▁▁▃▁▄▄▄▁▄▄▄▄▄▄▃▅▆███ █
  28.9 ns      Histogram: log(frequency) by time      30.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels._α_αp_asymptotic(v, x) setup=(v = rand()+20.0; x=rand()+50.0)
BenchmarkTools.Trial: 10000 samples with 986 evaluations.
 Range (min … max):  51.471 ns … 68.725 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     51.611 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   51.777 ns ±  0.811 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▇█▄▁                 ▁▁                                    ▁
  ██████▇▆▅▅▃▅▄▃▁▆▅▃▁▁▄▆███▇█▇▆▇▅▆▆▆▅▅▄▄▄▄▅▆▄▆▆▆▅▅▅▅▅▅▅▃▃▃▁▄▄ █
  51.5 ns      Histogram: log(frequency) by time      55.4 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels._α_αp_asymptotic(v, x) setup=(v = rand()+20.0; x=rand()+30.0)
BenchmarkTools.Trial: 10000 samples with 939 evaluations.
 Range (min … max):  103.414 ns … 130.634 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     103.576 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   103.797 ns ±   0.918 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▅▇█▇▅▄▄▂                           ▂▂▃▂▂▁▁                   ▂
  ██████████▇▇▇▆▆▆▆▅▄▅▅▄▅▁▄▃▃▃▃▁▁▃▃▁▆█████████▇▆▇█▇█▇▅▅▅▆▅▅▄▆▄▅ █
  103 ns        Histogram: log(frequency) by time        106 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.


## this PR
julia> @benchmark Bessels._α_αp_asymptotic(v, x) setup=(v = rand()+20.0; x=rand()+150.0)
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  21.361 ns … 33.596 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     21.504 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.686 ns ±  0.999 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂█                                                          ▁
  ██▇▆▄▁▃▄▄▃▄▄▃▃▅▅▄▆▆▆▅▃▅▅▄▆▅▆▄▆▇█▇▅▆▆▆▄▃▄▃▄▃▃▄▄▄▁▄▅▄▃▅▄▃▄▅▅▆ █
  21.4 ns      Histogram: log(frequency) by time      27.4 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels._α_αp_asymptotic(v, x) setup=(v = rand()+20.0; x=rand()+50.0)
BenchmarkTools.Trial: 10000 samples with 990 evaluations.
 Range (min … max):  42.863 ns … 67.257 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     43.449 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   43.565 ns ±  0.738 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

          ▄▂▄▃▆█▆▃▂▃▂▁▁                      ▁▁▁              ▂
  ▃▁▁▁▃▁▃▃█████████████▇▅▆▅▁▁▁▁▁▁▁▁▁▁▃▁▁▁▅▅▆▇████▇▇▆▅▅▆▆▆▄▅▅▆ █
  42.9 ns      Histogram: log(frequency) by time      45.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels._α_αp_asymptotic(v, x) setup=(v = rand()+20.0; x=rand()+30.0)
BenchmarkTools.Trial: 10000 samples with 946 evaluations.
 Range (min … max):  98.618 ns … 135.710 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     98.911 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   99.118 ns ±   1.207 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁▃▄ █▇▆                                                       
  ███▆███▆▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂ ▃
  98.6 ns         Histogram: frequency by time          102 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

There might be some more improvement possible but the vectorization is sensitive (see #44 for more discussion) to how the coefficients are used. There were several iterations I tried where vectorization was almost completely lost and benchmarks were over 3x greater....